requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
jinja2>=3.1.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.9
